Fixes exit codes + prints commands in convert.py
Note that (at least on OS X), a failing `colmap` command would return 256 which seems to result in a 0 error code of the script. Instead using custom error codes that can be used to distinguish failure cases easily when calling the script.

It's handy to have the commands logged as well. Note that they're only printed if `setLevel(INFO)` is called (which is also added in this commit).